Need to make .venv work, fix the errors, ui fix
